requirements.txt
speechrecognition
pyaudio
pyttsx3
openai
python-dotenv